fix notification preventing saving/loading
In Stupid Achievements v0.1.1, there's a bug that causes attempting to save your game while an achievement "task" is up to fail with the message "
Attempt to save pointer to unhandled type NativeStruct<Font>".

This commit fixes that by marking the offending mFont item as "transient". However, this also causes a VM abort when loading a save with a task up.

My knowledge about this VM abort is limited, but I do know the abort is related to the "isFinished" function. I've added a null check to this function, which fixes this VM abort but also now results in any currently running tasks to stop running when loading a game.